[Add] Prototype Synchronization and Multiple notes
Not everything works as intended yet but notes are created in the network.
