+++ Frontend/VIAProMa/Assets/Scripts/Multiplayer/Synchronizer/HaftnotizenSynchronizer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1a71024e7fecf93439bcdf43268e0cb5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
