+++ Frontend/VIAProMa/Assets/Scripts/Visualizations/Haftnotizen.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 07f85e8582ab88c4eab665980e6e071b
+guid: d98b11be57ac1cb44bb5879c21394fb7
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
