+++ Frontend/VIAProMa/Assets/Scripts/Visualizations/Haftnotizen/Haftnotizen.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cff04aa4759354741b24e4253a94e9ae
+guid: 4b6ca7abca177944c95e99ea57373160
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
